Updated Observer and Subject classes
Subjects now store observers in PriorityQueue instead of just list.
Observers now have a priority.

The reason is because PlayerController should always go first because it can transform player which will change it's instance.

I also added an extra argument to changeState and react for whether should react concurrently or not. This is just making a makeshift lock to make sure PlayerController finishes transforming before other changes happen to player. Only transformation does this situation happen but this is just to be safe.